requirements.txt
rich
aiosqlite
easy-pil
requests
python-dotenv